### 2017-08-29 342df10
Refactor distance internal names & improve fn signatures
