+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "finch"
-version = "0.1.2"
+version = "0.1.3"
 authors = ["Roderick Bovee & One Codex <roderick@onecodex.com>"]
 description = "An implementation of min-wise independent permutation locality sensitive hashing ('MinHashing') for genomic data and command-line utility for manipulation."
 keywords = ["minhash", "bioinformatics", "sketches"]
