### 2017-08-30 abc51c2
Add more traits to KmerCount
