### 2017-09-05 a52cbe9
Add Debug derives for (almost) all structs
