### 2017-09-07 825ef4d
Drop kmer length comparison in distance calc
