### 2017-09-14 b15864b
clarify types for rust/pull/44287
