### 2017-09-21 #378 @arfon
@biorelated @mgymrek would either of you be willing to edit this submission for JOSS?
