### 2017-09-22 f528a4d
Fix exit code 0 on return & start 0.1.4 branch
