### 2017-09-23 704ddcf
Add numValidKmers to sketches. Closes #7
