### 2017-09-28 e7d5c7c
Add `mash_stream` implementation
