+++ Cargo.toml
@@ -12,7 +12,7 @@ readme = "./README.md"
 [dependencies]
 clap = "~2.26.0"
 murmurhash3 = "~0.0.5"
-needletail = "~0.1.2"
+needletail = "~0.1.4"
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0.2"
