### 2017-10-31 fd7e52f
Fix bug where all reads would be filtered out when files are heavily adapter contaminated
