### 2017-11-13 #378 @boydgreenfield
@biorelated Thanks for offering to be the editor! I appreciate that this falls into your (probably quite limited) spare time, but have you had a chance to take a look yet? Please don't hesitate to ping me or @bovee if we can help in any way.
