### 2017-12-13 #505 @whedon
[REVIEW]: Finch: MinHashing for Sequencing Data with Abundance Calculation and Adaptive Filtering
**Submitting author:** @bovee (<a href="http://orcid.org/0000-0002-8819-9549">Roderick Boveee</a>)
**Repository:** <a href="https://github.com/onecodex/finch-rs" target ="_blank">https://github.com/onecodex/finch-rs</a>
**Version:** v0.1.4
