### 2017-12-15 #505 @HadrienG
The version in this review doesn't correspond to the latest version of the software. Unsure how to fix that here @biorelated. Can I simply edit the Whedon comment?

I will carry the review with the latest version 0.1.4 in the meantime
