### 2017-12-15 onecodex/finch-rs#9 @HadrienG
Linux executable
Hi!

I'm reviewing your [Joss submission](https://github.com/openjournals/joss-reviews/issues/505) and I was having troubles with the linux executable. It appears that  the `.gz` file available to download is actually a `tar` archive.

Instinctively I did

```bash
wget https://github.com/onecodex/finch-rs/releases/download/v0.1.4/finch-linux64-v0.1.4.gz
gzip -d finch-linux64-v0.1.4.gz
chmod +x finch-linux64-v0.1.4
./finch-linux64-v0.1.4
# -bash: ./finch-linux64-v0.1.4: cannot execute binary file
```

And indeed

```bash
file finch-linux64-v0.1.4
# finch-linux64-v0.1.4: POSIX tar archive (GNU)
```

I think it'd be less confusing for users to name the file `finch-linux64-v0.1.4.tar.gz`

Otherwise the software executes fine after unpacking! I may come back to you with others issues or suggestions later during the day 😃

Best,
Hadrien
