### 2017-12-15 #505 @arfon
> I will carry the review with the latest version 0.1.4 in the meantime

I've updated the comment at the top of the issue.
