### 2017-12-15 onecodex/finch-rs#9 @boydgreenfield
@HadrienG Thanks for catching that sloppy packaging! Just updated the binary to have the right extension.
