### 2017-12-15 f5a9858
Add appropriate extension to `-o` files if not present
