### 2017-12-19 onecodex/finch-rs#11 @HadrienG
Documentation
Hi again!

I'm more or less done testing the software itself and overall I must say it works well! I'd like to suggest a few improvements to the documentation if you don't mind:

- [x] Although the command-line options are well explained, I think it'd be a good improvement to have a quickstart section with one or two real examples, i.e. "How to compare my samples together?" or "How to quickly compare my samples to RefSeq?"

- [x] You provide sketches for RefSeq (this is great!) but it is not so straight-forward how to query those: `finch dist sketches/refseq_sketches_31_1000/*.sk` fails with `-bash: /Users/hadrien/.cargo/bin/finch: Argument list too long` on my computer. Having a walkthrough on how to do that would improve usability. I used `find` but not sure there isn't a better way.

- [x] There are not guidelines for contributing. Adding a small statement about where to report issues,  how to contribute to your software and eventually a Code of Conduct for people wising to contribute would be - in my opinion - great additions.

Best,
Hadrien.

<hr>

[Review ref for tracking](https://github.com/openjournals/joss-reviews/issues/505)
