### 2017-12-19 6140858
Create CODE_OF_CONDUCT.md
