### 2017-12-19 518f378
Add contribution section to README
