### 2017-12-19 onecodex/finch-rs#11 @bovee
Thanks! These are great suggestions.

I've gone ahead and added a small contribution section to the main README and selected a CODE_OF_CONDUCT (GitHub helpfully makes this very easy).

The first issue is fairly straight-forward, but for the second issue I think this may be a fundamental problem with how we've structured the databases. We should probably merge all of those into one file (because the `bash` problem is upstream of the code we have control over).
