### 2017-12-19 onecodex/finch-rs#11 @HadrienG
Thanks for your quick reply.

I do not think this is a fundamental problem and as you said what I stumbled upon is a `bash problem`. Of course it would be great to have a unified database or to allow `finch` to take a directory as input in mind for a future release.

For now I think potential users would be more than happy with a short tutorial in the README on how *you* use this databases.

I took the liberty of amending my issue with one more point.

Best,
Hadrien
