### 2017-12-20 onecodex/finch-rs#12 @HadrienG
JOSS paper
Hi!

Concerning the [JOSS paper](https://github.com/openjournals/joss-papers/blob/joss.00505/joss.00505/10.21105.joss.00505.pdf) it seems that:

- [ ] your affiliation is not complete, it would be nice to have the location of your company too.

- [ ] the MinHash reference reads `Broder, A.Z. n.d` probably because you don't have a `year` filed in `paper.bib` for this reference?

Best,
Hadrien

<hr>

[Review ref for tracking](https://github.com/openjournals/joss-reviews/issues/505)
