### 2018-01-02 e250a44
Update our address & some refs for JOSS paper
