+++ paper/paper.bib
@@ -1,8 +1,10 @@
 @inproceedings{Broder1997,
   doi = {10.1109/sequen.1997.666900},
   url = {https://doi.org/10.1109%2Fsequen.1997.666900},
+  year = {1997},
   publisher = {Institute of Electrical and Electronics Engineers (IEEE)},
-  author = {A.Z. Broder},
+  pages = {21--29},
+  author = {Andrei Z. Broder},
   title = {On the resemblance and containment of documents},
   booktitle = {Proceedings. Compression and Complexity of SEQUENCES 1997 (Cat. No.97TB100171)}
 }
@@ -38,7 +40,7 @@
   url = {https://doi.org/10.1016%2Fj.diagmicrobio.2016.09.023},
   year  = {2017},
   month = {jan},
-  publisher = {Elsevier {BV}},
+  publisher = {Elsevier},
   volume = {87},
   number = {1},
   pages = {11--16},
