### 2018-01-02 onecodex/finch-rs#12 @bovee
Thanks! Just added these two. As a note, there isn't an example format for affiliation addresses in the JOSS docs so I copied the format I see in a few previously submitted papers; please let me know if there's a better way.
