### 2018-01-10 03559dc
Change dist `query` selection, fix `info` stats, & speed up JSON loading. Closes #13.
