### 2018-01-10 1f3eabb
Single-file RefSeq examples. Closes part 2 of #11
