### 2018-01-10 d3e3460
Single-file RefSeq examples. Closes part 2 of #11
