### 2018-01-10 c7847db
Add example documentation of usage
