### 2018-01-16 e6568f9
Update README for clarity/style
