+++ README.md
@@ -8,34 +8,52 @@ This repository provides a library and command-line interface that reimplements 
 ## Getting Started ##
 
 ### Installation ###
-You may build Finch from source using Rust >= `1.15`, install via `cargo install finch` if you have Rust's Cargo package manager (see [rustup](https://www.rustup.rs) for fast Cargo installation instructions), or [download a prebuilt binary](https://github.com/onecodex/finch-rs/releases).
+You may build Finch from source using Rust >= `1.15`, install with `cargo install finch` if you have Rust's Cargo package manager (see [rustup](https://www.rustup.rs) for fast Cargo installation instructions), or [download a prebuilt binary](https://github.com/onecodex/finch-rs/releases).
 
 ### Example Usage ###
-Sketching sequencing files is the first step in most Mash pipelines:
+To get started, we first compute sketches for several FASTA or FASTQ files. These sketches are compact, sampled representations of the underlying genomic data, and what allow `finch` to rapidly estimate distances between datasets. Sketching files uses the `finch sketch` command:
+
 ```
 finch sketch example.fastq example2.fastq
 ```
-These sketches can be then be used with other Mash implementations (although see [this issue tracking interoperability](https://github.com/marbl/Mash/issues/27) and note other implementations may use a different seed value).
-All of Finch's commands can take either sketches or raw sequence files which they'll sketch on the fly so presketching sequencing files may not be necessary unless lots of downstream analyses are being performed.
+
+The resulting sketch files (`example.fastq.sk` and `example2.fastq.sk`) can then be used with other `finch` commands (as well as with other MinHash implementations<sup>1</sup>). Note that all of Finch's commands can take either sketches or raw sequence files. If passed the latter, `finch` will sketch the files on the fly. Sketches generated on the fly are **not saved**, however, so you should call `finch sketch` if you plan to use the sketch multiple times.
 
 Once sketched, multiple sequencing runs can be compared to determine how similar they are:
+
 ```
 finch dist example.fastq.sk example2.fastq.sk
 ```
-This will print a JSON object with some statistics including how much of the second file is found in the first file (`containment`) and how similar they are (`jaccard` index and `mashDistance`):
-```
-[{"containment":0.0,"jaccard":0.0,"mashDistance":1.0,"commonHashes":0,"totalHashes":1000,"query":"example2.fastq","reference":"example.fastq"}]
-```
-In this case, these files have nothing in common at the depth we're looking at!
-Resketching with a higher `--n-hashes` parameter may allow comparions of more dissimilar sequencing data like these.
 
-Finding the most similar relatives of one of these files in a RefSeq database may be helpful too (see *Example Data* below for links to pre-sketched RefSeq databases):
-```
-finch dist -q example.fastq -d 0.2 ./refseq_sketches_21_1000.sk ./example.fastq
-```
-A maximum distance of 0.2 is set here to filter out unrelated genomes (a distance of 0 would be an identical genome); setting a maximum is a good idea otherwise distances to *all* of RefSeq are returned.
+This will print results (in JSON) with some key distance statistics, including `containment` and `jaccard` similarity scores and a `mashDistance` distance estimate:
 
-**Note**: _More details on all of these commands and their parameters can be obtained with, e.g. `finch dist --help`._
+
+```json
+[
+  {
+    "commonHashes": 30,
+    "containment": 0.03,
+    "jaccard": 0.015228426395939087,
+    "mashDistance": 0.1669789474914277,
+    "query": "example2.fastq",
+    "reference": "example1.fastq",
+    "totalHashes": 1000
+  }
+]
+```
+
[…]
