### 2018-01-17 onecodex/finch-rs#11 @bovee
Hi Hadrien,

Sorry for the delay on closing this out (holidays!), but we added a "quick start" to our README and merged together the RefSeq sketches into single files (to fix the glob issue with Bash) in pull request #14, so I think this is complete now. Please let me know if there's anything I missed or if there's anything else you'd like with this.

Best,
Roderick
