### 2018-01-19 onecodex/finch-rs#11 @HadrienG
Hi!

No worries about the delay. I do not think I have other issues to raise. I'll test again with some example data with `0.1.5` and if it's runs fine I feel comfortable recommending for publication.

Best,
Hadrien
