### 2018-01-19 #505 @HadrienG
@arfon It's going pretty well. Vacations have slowed things a little but my last issue is now closed. I'll test the new release and if everything goes fine I fell comfortable recommending for publication
