### 2018-01-25 #505 @arfon
> @arfon Kindly review this for any final checks before consideration and publication.

Mostly this looks fine. I would encourage the authors to consider adding a little more detail to their paper as per our [submission guidelines](http://joss.theoj.org/about#paper_structure):

> - A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
> - A clear statement of need that illustrates the purpose of the software
> - A list of key references including a link to the software archive
> Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it
