### 2018-01-30 8d4441d
Update paper.md and paper.bib
