### 2018-01-30 29cbdb7
Update paper.md
Remove some commas and other final touch-ups
