### 2018-01-30 #505 @bovee
@arfon Thanks for the suggestions; we've expanded the `paper.md` to address those concerns.

@biorelated Please let me know if there are any remaining concerns or if there's anything else to do. Thanks!

More detailed notes:

> - A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience

The opening paragraph has been expanded to describe what MinHashing is used for and what Finch adds to traditional MinHashing schemes.

> - A clear statement of need that illustrates the purpose of the software

The second paragraph contains an expanded description of how MinHash works that motivates a discussion of the problems that Finch solves in the third paragraph.

> - A list of key references including a link to the software archive

We've added a link to the GitHub software repository.

> - Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it

The last paragraph now contains a description of how we are using Finch in our web platform.
