### 2018-01-31 #505 @biorelated
@bovee: Thank you for the submission and for addressing all the issues and concerns
@arfon:  Could we kindly consider this work for publication?
