### 2018-02-01 #505 @bovee
@arfon - I just submitted the repo to Zenodo and the record's at https://zenodo.org/record/1164259 and the DOI should be `10.5281/zenodo.1164259` (once it's propagated). Thanks!
